Исправлены мелкие ошибки, улучшена стабильность приложения

+++ apps/backend/src/main/java/io/github/artsobol/kurkod/web/domain/refreshtoken/repository/RefreshTokenRepository.java
@@ -8,5 +8,5 @@ import java.util.Optional;
 public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Integer> {
     Optional<RefreshToken> findByToken(String token);
 
-    Optional<RefreshToken> findByUserId(int userId);
+    Optional<RefreshToken> findByUserId(long userId);
 }
